lab1/requirements.txt
python-telegram-bot>=20.6,<21
python-dotenv>=1.0.1
ujson>=5.10.0
